backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bitarray==3.8.0
certifi==2026.2.25
ckzg==2.1.7
click==8.3.1
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==6.0.0b1
fastapi==0.135.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ollama==0.6.1
parsimonious==0.10.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
regex==2026.3.32
rlp==4.1.0
starlette==1.0.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
